+++ CHANGELOG.md
@@ -29,6 +29,7 @@
 - Expose metrics for the `AsyncPool`. ([#4538](https://github.com/getsentry/relay/pull/4538))
 - Expose service utilization metrics through the internal relay metric endpoint. ([#4543](https://github.com/getsentry/relay/pull/4543))
 - Always set observed time for OTel logs in Relay. ([#4559](https://github.com/getsentry/relay/pull/4559))
+- Switch the processor and store to `async`. ([#4552](https://github.com/getsentry/relay/pull/4552))
 
 ## 25.2.0